Version 1.1
Changed the noise control to not display a line but to allow for the control of the height of the extremities of a non-displayed line. The noise control is now applied in real-time, and applied to all sheets sharing the same base sheet (that gets modified). Changed the undo/redo to take into account these. 
Adapted the moment&fourier calculation to have a separate function for the setup and for the computing.
Changed the disable zoom to reduce dataset, with theta being updated each time, and the dataset reduced before creating other charts.